correções
atualizando

+++ src/main/java/com/taskify/api/controller/UsuarioController.java
@@ -4,6 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +19,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-
 import com.taskify.api.model.Usuario;
 import com.taskify.api.repository.UsuarioRepository;
 
@@ -21,18 +26,16 @@ import com.taskify.api.repository.UsuarioRepository;
 @RequestMapping(value = "/usuarios")
 public class UsuarioController {
 
-
     @PostMapping
-    public Usuario cadastrarUsuario(@RequestBody Usuario usuario){
-        return usuarioRepository.save(usuario);
-
+    public ResponseEntity<Usuario> cadastrarUsuario(@RequestBody Usuario usuario) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(usuarioRepository.save(usuario));
     }
-    
+
     @PutMapping("/{id}")
-    public Usuario atualizarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario) {
+    public ResponseEntity<Usuario> atualizarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario) {
         Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
-        
-        if(usuarioExistente.isPresent()){
+
+        if (usuarioExistente.isPresent()) {
             Usuario usuarioObj = usuarioExistente.get();
 
             usuarioObj.setNome(usuario.getNome());
@@ -41,39 +44,53 @@ public class UsuarioController {
             usuarioObj.setSenha(usuario.getSenha());
             usuarioObj.setGenero(usuario.getGenero());
 
-            return usuarioRepository.save(usuarioObj);
+            return ResponseEntity.status(HttpStatus.OK).body(usuarioRepository.save(usuarioObj));
         }
 
-        return null;
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @GetMapping
-    public List<Usuario> ListarUsuarios(){
-        return usuarioRepository.findAll();
+    public ResponseEntity<Page<Usuario>> listarUsuarios(
+            @PageableDefault(size = 2, page = 1, sort = "email", direction = Direction.DESC) Pageable paginacao) {
+        return ResponseEntity.status(HttpStatus.OK).body(usuarioRepository.findAll(paginacao));
     }
 
-
     @GetMapping("/{id}")
-    public Optional <Usuario> obterUsuarioPeloId(@PathVariable("id") Long id){
-        return usuarioRepository.findById(id);
+    public ResponseEntity<Optional<Usuario>> obterUsuarioPeloId(@PathVariable("id") Long id) {
+        Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
+
+        if (usuarioExistente.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(usuarioExistente);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @DeleteMapping("/{id}")
-    public void deletarUsuarioPeloId(@PathVariable("id") Long id){
+    public ResponseEntity<Void> deletarUsuarioPeloId(@PathVariable("id") Long id) {
+        Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
+
+        if (usuarioExistente.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         usuarioRepository.deleteById(id);
+
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     @GetMapping("/email/{email}")
-    public Optional<Usuario> obterUsuarioPeloEmail(@PathVariable("email") String email){
+    public Optional<Usuario> obterUsuarioPeloEmail(@PathVariable("email") String email) {
         return usuarioRepository.findByEmail(email);
     }
 
     @GetMapping("/nome/{nome}")
-    public Optional<List<Usuario>> obterUsuarioPeloNome(@PathVariable("nome") String nome){
+    public Optional<List<Usuario>> obterUsuarioPeloNome(@PathVariable("nome") String nome) {
         return usuarioRepository.findByNome(nome);
     }
 
     @Autowired
     private UsuarioRepository usuarioRepository;
-    
+
 }
